[fix] 테스트 가능하도록 타입 수정

+++ order-service/src/test/java/com/shoonglogitics/orderservice/domain/delivery/domain/vo/DeliveryActualTest.java
@@ -11,7 +11,7 @@ class DeliveryActualTest {
 	@DisplayName("DeliveryActual 객체 생성 시 거리와 소요시간이 정상적으로 세팅되어야 한다")
 	void createDeliveryActual_ShouldSetAllFields() {
 		// Given
-		Long distance = 1000L;
+		Integer distance = 1000;
 		Integer duration = 15;
 
 		// When
@@ -27,7 +27,7 @@ class DeliveryActualTest {
 	@DisplayName("거리가 음수이면 예외가 발생해야 한다")
 	void createDeliveryActual_WithNegativeDistance_ShouldThrowException() {
 		// Given
-		Long distance = -100L;
+		Integer distance = -100;
 		Integer duration = 15;
 
 		// When / Then
@@ -40,7 +40,7 @@ class DeliveryActualTest {
 	@DisplayName("소요시간이 음수이면 예외가 발생해야 한다")
 	void createDeliveryActual_WithNegativeDuration_ShouldThrowException() {
 		// Given
-		Long distance = 1000L;
+		Integer distance = 1000;
 		Integer duration = -5;
 
 		// When / Then
